Trying to catch when to build and when to install.  These hacks are making less and less sense.

+++ utility/ci/travis/install.sh
@@ -73,13 +73,12 @@ fi
 
 ./b2 headers
 
-cd "$BOOST_ROOT"
-echo "using $TOOLSET : : $COMPILER ;" > ~/user-config.jam
-IFS=','
-for CXXLOCAL in $CXXSTD; do  (cd libs/config/test && ../../../b2 config_info_travis_install toolset=$TOOLSET cxxstd=$CXXLOCAL && ./config_info_travis && rm ./config_info_travis)  done
-unset IFS
-cd "$BOOST_ROOT"
-./bootstrap.sh
-echo "$(pwd)"
-echo "$(ls)"
-$BOOST_ROOT/b2 -j3 libs/$SELF/test toolset=$TOOLSET cxxstd=$CXXSTD
+if [[ ! -z $CXXSTD ]] ; then
+	cd "$BOOST_ROOT"
+	echo "using $TOOLSET : : $COMPILER ;" > ~/user-config.jam
+	IFS=','
+	for CXXLOCAL in $CXXSTD; do  (cd libs/config/test && ../../../b2 config_info_travis_install toolset=$TOOLSET cxxstd=$CXXLOCAL && ./config_info_travis && rm ./config_info_travis)  done
+	unset IFS
+
+	./b2 -j3 libs/$SELF/test toolset=$B2_TOOLSET cxxstd=$CXXSTD
+fi
